3799: Extract BrushRegressionTest
