+++ common/test/CMakeLists.txt
@@ -128,6 +128,7 @@ set(COMMON_REGRESSION_TEST_SOURCE
         "${COMMON_TEST_SOURCE_DIR}/IO/Md3ParserRegressionTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/IO/Quake3ShaderParserRegressionTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Model/BrushNodeRegressionTest.cpp"
+        "${COMMON_TEST_SOURCE_DIR}/Model/BrushRegressionTest.cpp"
 )
